requirements.txt
matplotlib
opencv-python>=4.8.0
Pillow
ultralytics>=8.0.0
numpy>=1.24.0
zipfile36
flask==3.1.0
fastapi
websocket
flask-socketio==5.5.1
eventlet
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0

--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
